Nuget /content renamed
/content folder has been renamed to /project to avoid installing the source in Visual Studio by default.
