+++ src/Features/Sitecore.Pathfinder.Markdown/packages.config
@@ -4,5 +4,5 @@
   <package id="Markdown" version="1.14.4" targetFramework="net452" />
   <package id="Microsoft.Framework.ConfigurationModel" version="1.0.0-beta4" targetFramework="net452" />
   <package id="Microsoft.Framework.ConfigurationModel.Interfaces" version="1.0.0-beta4" targetFramework="net452" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
