+++ src/Features/Sitecore.Pathfinder.Npm/packages.config
@@ -3,5 +3,5 @@
   <package id="JetBrains.ExternalAnnotations" version="10.0.15" targetFramework="net46" />
   <package id="Microsoft.Framework.ConfigurationModel.Interfaces" version="1.0.0-beta4" targetFramework="net46" />
   <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
