+++ src/Features/Sitecore.Pathfinder.NuGet/NuGetPackageBuilder.cs
@@ -25,7 +25,7 @@ namespace Sitecore.Pathfinder.NuGet
             sb.AppendLine("        <description>Generated by Sitecore Pathfinder</description>");
             sb.AppendLine("    </metadata>");
             sb.AppendLine("    <files>");
-            sb.AppendLine("        <file src=\"" + sourceFileName + "\" target=\"content\\sitecore.project\\exports.xml\" />");
+            sb.AppendLine("        <file src=\"" + sourceFileName + "\" target=\"project\\sitecore.project\\exports.xml\" />");
             sb.AppendLine("    </files>");
             sb.AppendLine("</package>");
