+++ src/Features/Sitecore.Pathfinder.NuGet/packages.config
@@ -4,5 +4,5 @@
   <package id="Microsoft.Framework.ConfigurationModel.Interfaces" version="1.0.0-beta4" targetFramework="net46" />
   <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net46" />
   <package id="NuGet.Core" version="2.10.1" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
