+++ src/Features/Sitecore.Pathfinder.PageHtml/packages.config
@@ -4,5 +4,5 @@
   <package id="Microsoft.AspNet.Razor" version="3.1.1" targetFramework="net46" />
   <package id="Microsoft.AspNet.WebPages" version="3.1.1" targetFramework="net46" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
