+++ src/Features/Sitecore.Pathfinder.Roslyn/packages.config
@@ -5,7 +5,7 @@
   <package id="Microsoft.CodeAnalysis.Common" version="1.1.1" targetFramework="net46" />
   <package id="Microsoft.CodeAnalysis.CSharp" version="1.1.1" targetFramework="net46" />
   <package id="Microsoft.CSharp" version="4.0.0" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
   <package id="System.Collections" version="4.0.0" targetFramework="net46" />
   <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net46" />
   <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net46" />
