+++ src/Features/Sitecore.Pathfinder.T4/packages.config
@@ -4,5 +4,5 @@
   <package id="Microsoft.Framework.ConfigurationModel.Interfaces" version="1.0.0-beta4" targetFramework="net46" />
   <package id="Mono.TextTemplating" version="1.0.0" targetFramework="net46" />
   <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
