+++ src/Features/Sitecore.Pathfinder.Unicorn/packages.config
@@ -5,6 +5,6 @@
   <package id="Rainbow.Core" version="1.0.1" targetFramework="net46" />
   <package id="Rainbow.Storage.Sc" version="1.0.1" targetFramework="net46" />
   <package id="Rainbow.Storage.Yaml" version="1.0.1" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
   <package id="Unicorn.Core" version="3.0.2" targetFramework="net46" />
 </packages>
