+++ src/Sitecore.Pathfinder.Console/files/project.noconfig/sitecore.project/Sitecore.nuspec
@@ -13,8 +13,9 @@
         </dependencies>
     </metadata>
     <files>
-        <file src="**" target="content" exclude="**\*.cs;scc.cmd;sitecore.filetemplates\**;sitecore.project\schemas\**;.vscode\**;node_modules\**;bin\**\Sitecore*.*;bin\**\System*.*;bin\**\Microsoft*.*;bin\**\Lucene.Net.dll;bin\**\Newtonsoft.Json.dll;.vs\**;obj\**;Components\**;packages\**;*.sln;*.suo;*.csproj;*.user;*.csproj.sitecore;packages.config;gulpfile.js;**\readme.md;web.config;App_Browsers\**;App_Config\**;App_Data\**;Areas\**;aspnet_client\**;bin\**;data\**;sitecore\**;temp\**;upload\**;**\ExperienceExplorer\**;Social\**;WebEdit.css;default.*;Global.asax;_._;**/ExperienceExplorerView.cshtml" />
-        <file src="$toolsDirectory$\files\extensions\**\*.dll" target="content\sitecore.tools\files\extensions" />
-        <file src="$toolsDirectory$\scconfig.json" target="content\sitecore.tools\scconfig.json" />
+        <file src="bin\**" target="lib\net40" exclude="bin\**\Sitecore*.*;bin\**\System*.*;bin\**\Microsoft*.*;bin\**\Lucene.Net.dll;bin\**\Newtonsoft.Json.dll;" />
+        <file src="**" target="project" exclude="**\*.cs;scc.cmd;sitecore.filetemplates\**;sitecore.project\schemas\**;.vscode\**;node_modules\**;bin\**;.vs\**;obj\**;Components\**;packages\**;*.sln;*.suo;*.csproj;*.user;*.csproj.sitecore;packages.config;gulpfile.js;**\readme.md;web.config;App_Browsers\**;App_Config\**;App_Data\**;Areas\**;aspnet_client\**;bin\**;data\**;sitecore\**;temp\**;upload\**;**\ExperienceExplorer\**;Social\**;WebEdit.css;default.*;Global.asax;_._;**/ExperienceExplorerView.cshtml" />
+        <file src="$toolsDirectory$\files\extensions\**\*.dll" target="project\sitecore.tools\files\extensions" />
+        <file src="$toolsDirectory$\scconfig.json" target="project\sitecore.tools\scconfig.json" />
     </files>
 </package>
