+++ src/Sitecore.Pathfinder.Console/files/project/sitecore.project/sitecore.nuspec
@@ -13,9 +13,10 @@
         </dependencies>
     </metadata>
     <files>
-        <file src="**" target="content" exclude="**\*.cs;scc.cmd;sitecore.filetemplates\**;sitecore.project\schemas\**;sitecore.project\extensions\**;.vscode\**;node_modules\**;bin\**\Sitecore*.*;bin\**\System*.*;bin\**\Microsoft*.*;bin\**\Lucene.Net.dll;bin\**\Newtonsoft.Json.dll;.vs\**;obj\**;Components\**;packages\**;*.sln;*.suo;*.csproj;*.csproj.user;*.csproj.sitecore;packages.config;gulpfile.js;**\readme.md;web.config" />
-        <file src="sitecore.project\extensions\**\*.dll" target="content\sitecore.project\extensions" />
-        <file src="$toolsDirectory$\files\extensions\**\*.dll" target="content\sitecore.tools\files\extensions" />
-        <file src="$toolsDirectory$\scconfig.json" target="content\sitecore.tools\scconfig.json" />
+        <file src="bin\**" target="lib\net40" exclude="bin\**\Sitecore*.*;bin\**\System*.*;bin\**\Microsoft*.*;bin\**\Lucene.Net.dll;bin\**\Newtonsoft.Json.dll;" />
+        <file src="**" target="project" exclude="**\*.cs;scc.cmd;sitecore.filetemplates\**;sitecore.project\schemas\**;sitecore.project\extensions\**;bin\**;.vscode\**;node_modules\**;.vs\**;packages\**;obj\**;Components\**;packages\**;*.sln;*.suo;*.csproj;*.csproj.user;*.csproj.sitecore;packages.config;gulpfile.js;**\readme.md;web.config" />
+        <file src="sitecore.project\extensions\**\*.dll" target="project\sitecore.project\extensions" />
+        <file src="$toolsDirectory$\files\extensions\**\*.dll" target="project\sitecore.tools\files\extensions" />
+        <file src="$toolsDirectory$\scconfig.json" target="project\sitecore.tools\scconfig.json" />
     </files>
 </package>
