+++ src/Sitecore.Pathfinder.Console/packages.config
@@ -6,6 +6,6 @@
   <package id="Microsoft.Framework.ConfigurationModel.Json" version="1.0.0-beta4" targetFramework="net46" />
   <package id="Microsoft.Framework.ConfigurationModel.Xml" version="1.0.0-beta4" targetFramework="net46" />
   <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
   <package id="System.Management.Automation" version="6.1.7601.17515" targetFramework="net46" />
 </packages>
