+++ src/Sitecore.Pathfinder.Core/packages.config
@@ -6,5 +6,5 @@
   <package id="Microsoft.Framework.ConfigurationModel.Json" version="1.0.0-beta4" targetFramework="net45" />
   <package id="Microsoft.Framework.ConfigurationModel.Xml" version="1.0.0-beta4" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
