+++ src/Sitecore.Pathfinder.Server/packages.config
@@ -8,5 +8,5 @@
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
   <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net46" />
-  <package id="ResharperCodeContractNullability" version="1.0.4" targetFramework="net46" />
+  <package id="ResharperCodeContractNullability" version="1.0.5" targetFramework="net46" />
 </packages>
